inferno-comics-recog/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
numpy==2.0.2
opencv_python==4.12.0.88
Requests==2.32.4
tqdm==4.67.1
waitress==3.0.2
Werkzeug==3.1.3
psutil==7.0.0
PyYAML==6.0.2